# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CEO Tracker",
    "description": "Track CEO speeches and news for company mentions and sentiment",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/timeline": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Timeline",
        "description": "Get chronological feed of mentions.",
        "operationId": "get_timeline_api_timeline_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/sentiment/{sentiment}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Sentiment Companies",
        "description": "Get companies mentioned with specific sentiment.",
        "operationId": "get_sentiment_companies_api_dashboard_sentiment__sentiment__get",
        "parameters": [
          {
            "name": "sentiment",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sentiment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}